+++ shared/schemas/priceType/priceTypeBase.schema.ts
@@ -4,8 +4,8 @@ export const priceTypeBase = z.object({
   id: z.string().uuid(),
   type: z.string().min(1),
   size: z.string().min(1),
-  pixPrice: z.number().nonnegative(),
-  swilePrice: z.number().nonnegative(),
+  pixPrice: z.number().int().nonnegative(),
+  swilePrice: z.number().int().nonnegative(),
   deletedAt: z.date().nullable(),
 })
